Move stub relay hosts into testutils
Move these test helpers into testutils, instead of keeping them in a _test file.
